refs #23260 restructure AxisChanger
